ai-honeypot/functions/requirements.txt
azure-functions
azure-data-tables
azure-appconfiguration
azure-identity
